Home page defined settings

+++ b3-rest-api/helpers/B3_SettingsHelper.php
@@ -46,7 +46,7 @@ class B3_SettingsHelper {
             case 'show_on_front':
             case 'page_on_front':
             case 'page_for_posts':
-                return get_option( $option );
+                return (int) get_option( $option );
 
             default:
                 return get_bloginfo( $option );
